+++ opening-hours-syntax/src/rules/day.rs
@@ -1,6 +1,7 @@
-use std::convert::{TryFrom, TryInto};
-use std::fmt::Display;
-use std::ops::{Deref, DerefMut, RangeInclusive};
+use alloc::vec::Vec;
+use core::convert::{TryFrom, TryInto};
+use core::fmt::Display;
+use core::ops::{Deref, DerefMut, RangeInclusive};
 
 use chrono::prelude::Datelike;
 use chrono::{Duration, NaiveDate};
@@ -48,7 +49,19 @@ impl DaySelector {
             && self.weekday.is_empty()
     }
 
-    pub(crate) fn display(&self, f: &mut std::fmt::Formatter<'_>, force: bool) -> std::fmt::Result {
+    /// Format this day selector into given formatter.
+    ///
+    /// If `force` is set to true, this is guaranteed to yield a non-empty string by adding "Mo-Su"
+    /// as fallback.
+    pub(crate) fn display(
+        &self,
+        f: &mut core::fmt::Formatter<'_>,
+        force: bool,
+    ) -> core::fmt::Result {
+        if force && self.is_empty() {
+            return write!(f, "Mo-Su");
+        }
+
         if !(self.year.is_empty() && self.monthday.is_empty() && self.week.is_empty()) {
             write_selector(f, &self.year)?;
             write_selector(f, &self.monthday)?;
@@ -67,16 +80,12 @@ impl DaySelector {
             }
         }
 
-        if force && self.weekday.is_empty() {
-            write!(f, "Mo-Su ")
-        } else {
-            write_selector(f, &self.weekday)
-        }
+        write_selector(f, &self.weekday)
     }
 }
 
 impl Display for DaySelector {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         self.display(f, false)
     }
 }
@@ -108,7 +117,7 @@ pub struct YearRange {
 }
 
 impl Display for YearRange {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         write!(f, "{}", self.range.start().deref())?;
 
         if self.range.start() != self.range.end() {
@@ -138,7 +147,7 @@ pub enum MonthdayRange {
 }
 
 impl Display for MonthdayRange {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         match self {
             Self::Month { range, year } => {
                 if let Some(year) = year {
@@ -199,7 +208,7 @@ impl Date {
 }
 
 impl Display for Date {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         match self {
             Date::Fixed { year, month, day } => {
                 if let Some(year) = year {
@@ -259,7 +268,7 @@ impl DateOffset {
 }
 
 impl Display for DateOffset {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         write!(f, "{}", self.wday_offset)?;
         write_days_offset(f, self.day_offset)?;
         Ok(())
@@ -283,7 +292,7 @@ impl Default for WeekDayOffset {
 }
 
 impl Display for WeekDayOffset {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         match self {
             Self::None => {}
             Self::Next(wday) => write!(f, "+{}", wday_str(*wday))?,
@@ -311,7 +320,7 @@ pub enum WeekDayRange {
 }
 
 impl Display for WeekDayRange {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         match self {
             Self::Fixed { range, offset, nth_from_start, nth_from_end } => {
                 write!(f, "{}", wday_str(*range.start()))?;
@@ -367,7 +376,7 @@ pub enum HolidayKind {
 }
 
 impl Display for HolidayKind {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         match self {
             Self::Public => write!(f, "PH"),
             Self::School => write!(f, "SH"),
@@ -403,7 +412,7 @@ pub struct WeekRange {
 }
 
 impl Display for WeekRange {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         if self.range.start() == self.range.end() && self.step == 1 {
             return write!(f, "{:02}", **self.range.start());
         }
@@ -490,7 +499,7 @@ impl Month {
 }
 
 impl Display for Month {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         write!(f, "{}", &self.as_str()[..3])
     }
 }